Refactored MultiChatSpigot (LARGE)
